Adding generic family to the font-family Property
adding a generic family since there's no guarantee that a specific font is installed on the computer or can be downloaded using a @font-face at-rule. Also add an extra font-type for the browser the users is using to have a fall back in case the font-type is not present, and if none of the two isn't present, the browser can pick from any of the 'sans-serif' generic family and not serif, cursive, monospace and fantasy family.